requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
python-multipart

google-generativeai
huggingface-hub

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq

faiss-cpu

pillow
pdf2image

vertexai
google-cloud-aiplatform
sentence-transformers  

qdrant-client
